Implement Task 189 and 190: C89Emitter and CVariableAllocator
- Extended Platform Abstraction Layer (PAL) with file I/O primitives.
- Implemented `C89Emitter` with 4KB buffering and indentation support.
- Implemented `CVariableAllocator` with C89 keyword avoidance (binary search),
  truncation to 31 chars, and uniquification.
- Integrated `C89Emitter` into `CompilationUnit::generateCode` using stack allocation.
- Updated `DESIGN.md` and added `C89_Codegen.md` for documentation.
- Verified with comprehensive unit tests in new Batch 22 and 23.
- All tests pass, adhering to C++98 and <16MB memory constraints.

+++ src/bootstrap/utils.cpp
@@ -49,13 +49,13 @@ void simple_itoa(long value, char* buffer, size_t buffer_size) {
         n = -n;
     }
 
-    while (n != 0 && i < (int)buffer_size - (is_negative ? 2 : 1)) {
-        int rem = n % 10;
+    while (n != 0 && (size_t)i < buffer_size - (is_negative ? 2 : 1)) {
+        int rem = (int)(n % 10);
         buffer[i++] = rem + '0';
         n = n / 10;
     }
 
-    if (is_negative && i < (int)buffer_size - 1) {
+    if (is_negative && (size_t)i < buffer_size - 1) {
         buffer[i++] = '-';
     }
